Android Attendance App
Attendance Management System app is software developed for daily student
attendance in schools, colleges and institutes. It facilitates to access the
attendance information of a particular student in a particular class. The
information is sorted by the operators, which will be provided by the teacher fora
particular class. This system will also help in evaluating attendance
eligibilitycriteria of a student.
Student Attendance Management System app deal with the maintenance of the
student’s attendance details. It is generates the attendance of the student on
basis of presence in class. It is maintaining daily basis of attendance, the staff
will be provide with the separate username and password to make
student Attendance. The staff handling the particular subject to responsible to
make the attendance for all students. Only if the student presents the particular
date, the attendance will be calculated. The student attendance report based on
monthly and consolidate will be generated.